stdeb.util: Use explicit \t
Line 1312 is *supposed* to replace instances of 8 spaces with tabs, but
sometimes it misses some.  Trying to fix this could be problematic.

*Much* better to accept that `make` require tabs, `debian/rules` is a
`Makefile`, therefore, we should use tabs regardless of the Python coding
standards.

Thankfully, Python gives us a literal character we can insert to do
exactly this, so we can keep our space-indented code the way it is.

Ugly code that works beats pretty code that doesn't!